Re desigining the scope of the knowledge Recalled class
This class should be responsible for for initializing the application, managing hooks and their states, but things that
occur on any of the data models should occur their rather in this location. Anything pertaining to hooks likely need
there own HookManagement class which should initialize some sort of event listener and maintain a singleton state.